+++ packages/ssr/src/nextjs/DataProvider/createPersistedStore.ts
@@ -1,0 +1,8 @@
+import createPersistedStoreClient from './createPersistedStoreClient.js';
+import createPersistedStoreServer from './createPersistedStoreServer.js';
+
+const createPersistedStore =
+  typeof window === 'undefined' ?
+    createPersistedStoreServer
+  : createPersistedStoreClient;
+export default createPersistedStore;
